eunjong_update_day1

+++ Order/src/main/java/hanwhadeliverysystemteam/domain/MenuCancelled.java
@@ -11,11 +11,12 @@ public class MenuCancelled extends AbstractEvent {
 
     private Long id;
     private String menuName;
-    private Integer orderId;
+    private Long orderId;
     private String address;
     private Integer qty;
     private String orderStatus;
     private Double price;
+    
 
     public MenuCancelled(Order aggregate) {
         super(aggregate);
