+++ Order/src/main/java/hanwhadeliverysystemteam/domain/MenuOrdered.java
@@ -11,11 +11,12 @@ public class MenuOrdered extends AbstractEvent {
 
     private Long id;
     private String menuName;
-    private Integer orderId;
+    private Long orderId;
     private String address;
     private Integer qty;
     private String orderStatus;
     private Double price;
+    
 
     public MenuOrdered(Order aggregate) {
         super(aggregate);
